Improved Folder Structure
Structure of folders improved
