--- Animation.h
@@ -1,9 +1,0 @@
-#pragma once
-class Animation
-{
-	public:
-		Animation();
-		~Animation();
-	private:
-};
-
