+++ GameBackend/Genjutsu.cpp
@@ -1,0 +1,13 @@
+#include "Genjutsu.h"
+
+
+
+Genjutsu::Genjutsu(std::string s, int dmg, int hCost, int cCost) : Technique(s, dmg, hCost, cCost)
+{
+	
+}
+
+
+Genjutsu::~Genjutsu()
+{
+}
