+++ GameBackend/Ninjutsu.cpp
@@ -1,0 +1,12 @@
+#include "Ninjutsu.h"
+
+
+
+Ninjutsu::Ninjutsu(std::string s, int dmg, int hCost, int cCost) : Technique(s, dmg, hCost, cCost)
+{
+}
+
+
+Ninjutsu::~Ninjutsu()
+{
+}
