+++ GameBackend/Taijutsu.cpp
@@ -1,0 +1,13 @@
+#include "Taijutsu.h"
+
+
+
+Taijutsu::Taijutsu(std::string s, int dmg, int hCost, int cCost) : Technique(s, dmg, hCost, cCost)
+{ 
+	
+}
+
+
+Taijutsu::~Taijutsu()
+{
+}
