--- Ninjutsu.cpp
@@ -1,12 +1,0 @@
-#include "Ninjutsu.h"
-
-
-
-Ninjutsu::Ninjutsu(std::string s, int dmg, int hCost, int cCost) : Technique(s, dmg, hCost, cCost)
-{
-}
-
-
-Ninjutsu::~Ninjutsu()
-{
-}
