--- Person.cpp
@@ -1,24 +1,0 @@
-#include "Person.h"
-
-
-
-Person::Person(std::string n, int hp, int ps, int l) : name(n), hitPoints(hp), physicalStrength(ps), level(l)
-{
-}
-
-Person::~Person()
-{
-}
-void Person::setName(std::string n)
-{
-	name = n;
-}
-std::string Person::getName()
-{
-	return name;
-}
-
-int Person::getHP()
-{
-	return hitPoints;
-}
